Rawr3.BossHandler: Worked in the last of the options that have already been programmed into the boss handler onto the pane. What you see is what there is. If you want more options to be considered, post in the forums on the designated threads or make a feature request in the issue tracker.
git-tfs-id: [https://tfs.codeplex.com/tfs/TFS03]$/Rawr;C52088

+++ Rawr.DPSWarr/Skills.Shared.cs
@@ -399,7 +399,7 @@ namespace Rawr.DPSWarr.Skills
             ReqMeleeWeap = true;
             ReqMeleeRange = true;
 #if RAWR3 || SILVERLIGHT
-            Targets += (BossOpts.MultiTargs ? (BossOpts.MaxNumTargets - 1f) : 0f);
+            Targets += (BossOpts.MultiTargs ? ((float)BossOpts.MaxNumTargets - 1f) : 0f);
 #else
             Targets += (CalcOpts.MultipleTargets ? (CalcOpts.MultipleTargetsMax - 1f) : 0f);
 #endif
